+++ MapboxNavigation/CarPlayManager.swift
@@ -551,8 +551,7 @@ extension CarPlayManager: CPMapTemplateDelegate {
         let mapView = carPlayMapViewController.mapView
         let route = routeChoice.userInfo as! Route
         
-        let distanceFormatter = DistanceFormatter()
-        let estimates = CPTravelEstimates(distanceRemaining: distanceFormatter.measurement(of: route.distance),
+        let estimates = CPTravelEstimates(distanceRemaining: Measurement(distance: route.distance).localized(),
                                           timeRemaining: route.expectedTravelTime)
         mapTemplate.updateEstimates(estimates, for: trip)